Expand `TaskOutput::Return` values in the turbo-tasks macros
Previously, we'd return types like:

```
<Vc<AssetIdent> as TaskOutput>::Return
```

which is equivalent to the much simpler:

```
Vc<AssetIdent>
```

It's nice because it's easier to implement in the macros, and it's more
correct, because it uses the type system to verify the type of `Vc` and
`Result` are what we expect. However, it makes the rustdoc output hard
to read.

This attempts to the expansion of `TaskOutput::Return` inside the macro
to generate more readable documentation, at the cost of some
correctness.