fix(Turborepo): Handle new packages in lockfile comparisons (#8127)
### Description

When comparing lockfiles between revisions to identify what packages
have been affected, we treated missing packages in the previous lockfile
(added in the new lockfile) as errors, and fell back to considering all
packages as affected. This change adds a boolean to determine what the
behavior should be when we can't find a package that we expect. When
building the package graph, it continues to be an error, as well as when
tracing the dependencies of a package we've already found in the
lockfile. However, for the first round of packages that we look for in a
previous lockfile, we ignore missing packages. This allows us to more
gracefully handle comparisons between commits that add new packages.

### Testing Instructions

Added an integration test inspired by the repro from the linked issue.

Fixes #8125

---------

Co-authored-by: Greg Soltis <Greg Soltis>