feat(turborepo): new ui + watch mode (#7962)
### Description

Integrates new UI with watch mode. Pulls out the UI handle and sender
outside the `Visitor` and `Run` struct, so it can be owned by the
`WatchClient`. Also adds an `Event` for updating the task names, so on
rediscovery we can keep using the same UI thread but just update the
task names.

### Testing Instructions

Give it a shot!


Closes TURBO-2801

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>