Update getting started docs to disambiguate gitignore pattern
Update getting started docs to disambiguate gitignore pattern. The existing docs leads you to believe that the glob pattern is `+ .turbo` which *includes* turbo files in your git commits, instead of `.turbo`, which hides them. Removed the `+` to remove this ambiguity (in the Docs UI it isn't that it was referencing a "diff", and that the `+` wasn't part of the pattern).