docs: update links to logs in turbo types (#8403)
### Description

<!--
  ✍️ Write a short summary of your work.
  If necessary, include relevant screenshots.
-->

I noticed that quite a few of the links in the JSON schema's reference
are outdated. I have only updated links to fields that still exist, not
removed `dotEnv` and `globalDotEnv`.

Some of the links are changed only in capitalization, this is because
Edge (the only browser I tested in) did not recognize the anchors if the
capitalization was not identical to the anchors.

After updating the JSDocs, I generated a new version of the JSON schema
in `docs/public`. Please verify that this is now correct.

---------

Co-authored-by: Sindre Haugland Bøyum <sindre.haugland.boyum@dnb.no>
Co-authored-by: Anthony Shew <anthony.shew@vercel.com>