Add/move tests for Vc generics (#8843)
## Description

- Move the tests out of `all_in_one` so that we can see exactly which
parts are failing more easily.
- Add a few more tests. **The newly added `test_read_ref_round_trip`
fails** (and is ignored), indicating a bug in the generics
implementation!!! This is fixed in #8845.

## Why does `test_read_ref_round_trip` fail?

The theory is that Vcs are supposed to be stored as `<<T as
VcValueType>::Read as VcRead<T>>::Repr`. However, it looks like due to
an oversight, `ReadRef::cell` is trying to store the value as `T`.

## Why add these tests at all?

The plan (as of yesterday) is to remove support for generics, but I'd
feel more confident if I can fix the casting issues in my local Vc PRs
before I do that.

These tests should help me understand what's happening and debug things.

## Testing Instructions

```
cargo nextest r -p turbo-tasks -p turbo-tasks-memory
```