feat(turborepo): log reason why all packages were considered changed (#8872)
The ultimate goal for me is to figure out why I'm seeing _all_ packages
invalidated when running with `--filter='[sha]'` . In a lot of cases, we
return a sigil for `PackageChanges::All`. I'm not sure how to bubble
this up with higher fidelity information in a good way, so this PR
settles for a more granular "reason" why all packages changed and
includes it in debug logging. This should help a little bit with
debugging.

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>