feat: implement a query interface for turborepo (#8977)
### Description

Adds a GraphQL query interface to turborepo via the `turbo query`
command. If you pass it a query string, then it directly runs the query.
If you pass it a file, it reads the file and runs the query. If you
don't pass a query, then it opens up a server with GraphiQL (we can ITG
this interface).


### Testing Instructions
Added tests to `affected.t` and `command-query.t`

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>
Co-authored-by: Thomas Knickman <tom.knickman@vercel.com>