fix(filter): validate package name on full graph (#9102)
### Description

Fixes #9096

The issue was that the validation was run against an already filtered
set of packages if another selector was provided i.e. `--filter
foo[commit]` the name filter `foo` would run against only packages that
were changed since `commit`. This meant that `foo` would be incorrectly
reported as not existing.

This PR changes so the package existence check is run against all
packages in the graph instead of an already filtered set.

### Testing Instructions

Added red->green unit test


Verify that this PR correctly handles the repro provided in #9096 
```
[1 olszewski@chriss-mbp] /tmp/turbo-filter-break $ turbo_dev --skip-infer run lint '--filter=...@repo/eslint-config[HEAD~1]' 
turbo 2.1.1

• Packages in scope: 
• Running lint in 0 packages
• Remote caching disabled

No tasks were executed as part of this run.

 Tasks:    0 successful, 0 total
Cached:    0 cached, 0 total
  Time:    75ms
```