refactor(query): make immediate dependencies default (#9114)
### Description

We have a `dependencies` field on our `Package` type in `turbo query`.
This returned all the transitive dependencies, i.e. the dependencies of
the package's dependencies and their dependencies, and so on. This is a
little confusing, so now we have explicit `directDependencies`,
`indirectDependencies` and `allDependencies` fields.

### Testing Instructions

Tests in `command-query.t` are updated.