docs: add -w flag to pnpm example (#9160)
It's suggested to install this into the repository root. If you do this,
you must use `--workspace-root` flag or you will get this error:


>  ERR_PNPM_ADDING_TO_ROOT  Running this command will add the dependency
to the workspace root, which might not be what you want - if you really
meant it, make it explicit by running this command again with the -w
flag (or --workspace-root). If you don't want to see this warning
anymore, you may set the ignore-workspace-root-check setting to true.