feat: turbo-trace (#9134)
### Description

A long long time ago we had a
[node-file-trace](https://github.com/vercel/nft) implementation in the
repository. Unfortunately this was tied to turbopack and the turbo-tasks
code. Since that's now in a different repository, we're gonna implement
our own version, because that's easier than coordinating multi-repo
dependencies.

This implementation uses `oxc_resolver` and `swc` to resolve
dependencies and parse files.

This PR hooks it up to `turbo query` but we'll probably have other uses
later.

### Testing Instructions

Adds some tests in `turbo-trace.t` using `query`

---------

Co-authored-by: Turbobot <turbobot@vercel.com>