docs: Task outputs clarifications (#9185)
To avoid the impression that cache is disabled when Task `outputs` key
is ommitted.

`API reference` was more accurate for that behavior, than `Crafting your
repository`.

### Description

I'm new to Turborepo, and was reading [Crafting your
repository](https://turbo.build/repo/docs/crafting-your-repository). Two
sections lead me to conclude that when there is no `outputs` key in a
Task, then caching is disabled. I think the wording in `API reference`
is better.

We have to account for Logs output even if there are no file outputs
(which is really nice for linters).

### Testing Instructions

* Review the modified documentations.
* Take into account that I'm not a native english speaker (you probably
notice already ;) )