fix: unnecessary empty cache logs (#9301)
### Description

In #9236 I added a check if the task output globs are empty. Turns out
we need to check if the output globs only has a single entry for the log
file.

Fixes #9291

### Testing Instructions

`basic_monorepo` has both a task that has globs and no outputs and a
task that has empty globs and no outputs. Note that we warn for one and
not the other.