fix: use oxc-resolver hot fix for symlinks (#9302)
### Description

There's a [bug in
oxc-resolver](https://github.com/oxc-project/oxc-resolver/issues/263)
for symlinks that basically breaks tsconfig resolution with pnpm (in a
specific scenario for extending a tsconfig that's a workspace
dependency). Until that's fixed, this PR switches us to a fork that has
a very dumb hot fix

We also add tsconfig loading, either via a parameter or via walking up
the path hierarchy for the closest tsconfig

### Testing Instructions

Added a test based off of [my
repro](https://github.com/nicholaslyang/oxc-repro). Validated that
doesn't work with main branch of oxc-resolver, but does work on my fork