feat(query): reverse tracing (#9319)
### Description

Implements an MVP of reverse tracing. Basically goes through all the
JavaScript/TypeScript files in the repo, checks which ones import the
reverse traced file, and returns those. Does this concurrently with a
`JoinSet`.

Also adds code to detect if a tsconfig is in scope for a file and if so,
add it to the resolver. That way custom aliases work.

### Testing Instructions

Added some tests including a monorepo setup with tsconfigs in the
packages that define aliases