feat(ui): Hide task list with `h` key. (#9350)
### Description

Toggle the visibility of the task list with an `h` hotkey.


https://github.com/user-attachments/assets/7a8b3ad1-5ead-47d0-9aaf-513acce6ad3c

Future note: I don't love that I'm stealing a vertical line from the
task list. Screen real estate is valuable. It's likely we could give the
`Task` header and footer instructions a faded style and get rid of the
horizontal bars. (I'd recommend doing this regardless of the result of
this PR.

### Testing Instructions

👀