fix: order tasks in order of interest, namely failed, successful, then cached (#9353)
Currently, when running tasks using the new UI, turbo inserts passed / failed tasks below cached tasks (in order of completion) meaning the stuff users probably want to see (failed things) end up at the bottom of the list or off screen. This change attempts to improve that by inserting those items in front of the list of cached tasks. That way cached tasks (likely the least interesting) end up last and preceded by successful tasks and then failed tasks. This assumes that users will more interested in failed tasks than successful ones but I am also open to ordering just the failed / successful task in order of completion instead.

In practice this means if running turbo in a large project with `--continue` where the majority of tasks are and will remain cached between runs, then the things that are 'interesting' will usually be close to the user's cursor.

It also 'fixes' the case where viewing a task that is completed jumps past the cached tasks and you have to traverse back through them all to get to the top again.