feat(mfe): add env var indicating mfe proxy (#9395)
### Description

MFE development tasks may want to know if there's a corresponding proxy
that is being run. Just checking if the proxy is running is error prone
since they're both persistent tasks and there's no consistent way to
prevent the check from running before the proxy server is listening on
the port.

We avoid needing to perform this check by setting an environment
variable for each development task if there's an associated proxy task
in the configuration. It is then up to the underlying task to alter it's
behavior depending on the presence of this variable.

### Testing Instructions

In a project with a MFE, run a development task and verify that
`TURBO_TASK_HAS_MFE_PROXY` is set to `true`