feat: Detect WSL in `turbo info`. (#9416)
### Description

We have had cases in the past where it would be nice to know if the user
is using WSL. This adds best effort detection for printing out if you're
using WSL.

Note that this isn't a 100% solution, according to [this superuser
answer](https://superuser.com/questions/1749781/how-can-i-check-if-the-environment-is-wsl-from-a-shell-script/1749811#1749811).
But, if you're doing something like is listed in the caveats...That's an
extreme edge case.