feat(tui): Enter interactive with `i` key. (#9479)
### Description

We've begun to embrace the letters on the keyboard a bit more for
manipulating the TUI.

This PR allows you to enter into interacting with a task using the `i`
key. Notably, it leaves `Enter` in place, despite the helper text
changing. This is to let muscle memory for existing users keep at it -
until we remove it in 3.0. Hopefully they see the helper text by then.

### Testing Instructions

Try it out!

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>