refactor(tui): Gain better aesthetic for helper text. (#9478)
### Description

Just wanting to spruce things up a bit.

- Left-aligned the text so that it bounces around less when changing
states. This is considered a generally good practice in UI design.
- Re-writing text chunks for simplicity

In the process of adjusting the text with a comma as a separator, I
realized I could refactor the logic for which text to render to be a
little easier to read. Isn't it great when the simplicity of the visual
design is reflected in the code?!

### Testing Instructions

You'll want to put the TUI into different states to see them all. Matrix
is:
- Task list open/closed
- Text selected/not selected
- Task being interacted with/ not interacted with

A sample of the spirit of the changes is below.

Before
![CleanShot 2024-11-20 at 23 05
58@2x](https://github.com/user-attachments/assets/e5ba1e5b-54e1-4cc6-9dee-20ac2e17d303)

After
![CleanShot 2024-11-20 at 23 06
38@2x](https://github.com/user-attachments/assets/50be8c06-fbfb-4e46-bce7-d50da8f83f9e)

---------

Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>