feat: update `eslint-config-turbo` to support ESLint Flat Config. (#9502)
### Description

Adds ESLint Flat Config Support to `eslint-config-turbo`. Users will be
able to use:
```
import turboConfig from "eslint-config-turbo/flat"

export default [
  ...turboConfig
]
```

This is meant to be backwards compatible with ESLint v8, leaving the
`extends` key in place for that purpose.

### Testing Instructions

Difficult to test this, so we'll be cutting a canary shortly after to
make sure.