chore(examples): Remove spare dependencies from basic example. (#9516)
### Description

I left some unneeded dependencies behind when I was upgrading this
example in #9501. This gets rid of them.

### Testing Instructions

I've verified that ESLint violations get reported as expected in my
editor and when running `turbo lint`. You can also clone with `npx
create-turbo@canary -e
https://github.com/vercel/turborepo/tree/shew-0f804/examples/basic
basic-eslint-9` to have a play with it.