chore(docs): enable hashlinking for System Environment Variables entries (#9532)
### Description

Before, we were only able to link people to the page itself, and having
them find the variable in the table itself.

Instead, let's put an `id` tag on things so we can give hashlinks to
users.

### Testing Instructions

This should be a no-op on the visual rendering of the page, as I've
pasted the rendered HTML from the page directly into the Markdown file.
The only change is the addition of the `id`s.