chore(pm): reduce PackageManager surface area (#9557)
### Description

A prefactor for cleaning up the `PackageManager` enum. If we want to
make this a richer data structure we need to remove some behaviors that
other crates might be depending on. This is just a small PR, the far
larger one will be removing direct pattern matching on `PackageManager`
variants.

### Testing Instructions

This is purely removal of code and moving of code. Relying on existing
unit tests